chore: extract event-emitter from core
