+++ packages/core/src/common/app-event-bus/app-event-bus.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { EventEmitter } from "../event-emitter";
+import { EventEmitter } from "@k8slens/event-emitter";
 import type { AppEvent } from "./event-bus";
 
 const appEventBusInjectable = getInjectable({
