+++ packages/core/src/extensions/common-api/event-bus.ts
@@ -6,7 +6,7 @@
 import appEventBusInjectable from "../../common/app-event-bus/app-event-bus.injectable";
 import { asLegacyGlobalForExtensionApi } from "../as-legacy-globals-for-extension-api/as-legacy-global-object-for-extension-api";
 import type { AppEvent } from "../../common/app-event-bus/event-bus";
-import type { EventEmitter, EventEmitterCallback, EventEmitterOptions } from "../../common/event-emitter";
+import type { EventEmitter, EventEmitterCallback, EventEmitterOptions } from "@k8slens/event-emitter";
 
 export type {
   AppEvent,
