+++ packages/open-lens/package.json
@@ -184,6 +184,7 @@
     "@k8slens/application-for-electron-main": "^6.5.0-alpha.3",
     "@k8slens/core": "^6.5.0-alpha.8",
     "@k8slens/ensure-binaries": "^6.5.0-alpha.4",
+    "@k8slens/event-emitter": "^1.0.0-alpha.1",
     "@k8slens/feature-core": "^6.5.0-alpha.3",
     "@k8slens/keyboard-shortcuts": "^1.0.0-alpha.3",
     "@k8slens/kubectl-versions": "^1.0.0-alpha.2",
