+++ packages/utility-features/event-emitter/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: "@k8slens/eslint-config/eslint",
+  parserOptions: {
+    project: "./tsconfig.json",
+  },
+};
