+++ packages/utility-features/event-emitter/.prettierrc
@@ -1,0 +1,1 @@
+"@k8slens/eslint-config/prettier"
