+++ packages/utility-features/event-emitter/index.ts
@@ -1,0 +1,1 @@
+export * from "./src/event-emitter";
