+++ packages/utility-features/event-emitter/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@k8slens/jest").monorepoPackageConfig(__dirname).configForReact;
