+++ packages/utility-features/event-emitter/webpack.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@k8slens/webpack").configForNode;
